% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfr.R
\name{cfr}
\alias{cfr}
\alias{cfr_ui}
\alias{cfr_server}
\title{CFR module}
\usage{
cfr_ui(id, full_screen = TRUE)

cfr_server(id, df)
}
\arguments{
\item{id}{Module id. Must be the same in both the UI and server function to link the two.}

\item{full_screen}{Add button to card to with the option to enter full screen mode?}

\item{df}{A \verb{<data.frame>} of daily cases and deaths to be passed to
\code{launch_module()}.
Must include the columns \code{date}, \code{cases}, and \code{deaths}, which specify the
daily cases and deaths reported during the outbreak.
The CFR module currently only supports aggregated incidence data and does
not support grouping variables.
Dates must be a continuous series and no values may be missing or \code{NA}.
See \code{cfr::cfr_rolling()} or \code{cfr::cfr_time_varying()} for more details on
the CFR functions.}
}
\value{
Creates a Shiny module to be launched by \code{launch_module()}.
}
\description{
Visualise the rolling or time-varying disease severity in the
form of the case fatality risk, while optionally correcting for delays in
reporting outcomes (deaths).
}
