% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{df_ll}
\alias{df_ll}
\title{Example Linelist Data}
\format{
a tibble dataframe
}
\usage{
df_ll
}
\description{
A 'linelist' is a (tidy) data format used in public health data collection
with each row representing an individual (patient, participant, etc) and
each column representing a variable associated with said individual.
}
\details{
\code{df_ll} is an example linelist dataset containing data for a fake measles
outbreak in Yemen. The data contains temporal, demographic, and geographic
information for each patient, as well as other medical indicators.
}
\examples{
df_ll
}
\keyword{datasets}
