% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{df_ll_ebola}
\alias{df_ll_ebola}
\title{Example Ebola Linelist Data}
\format{
a tibble dataframe
}
\usage{
df_ll_ebola
}
\description{
A 'linelist' is a (tidy) data format used in public health data collection
with each row representing an individual (patient, participant, etc) and
each column representing a variable associated with said individual.
}
\details{
\link[outbreaks:ebola_sim]{outbreaks::ebola_sim_clean} similated ebola outbreak in Freetown, Sierra Leone.
Ages and admin boundary pcodes were added by the epishiny author.
}
\examples{
df_ll
}
\keyword{datasets}
