% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{df_ll_measles}
\alias{df_ll_measles}
\title{Example Measles Linelist Data for Chad}
\format{
a tibble dataframe
}
\usage{
df_ll_measles
}
\description{
A 'linelist' is a (tidy) data format used in public health data collection
with each row representing an individual (patient, participant, etc) and
each column representing a variable associated with said individual.
}
\details{
\code{df_ll_measles} contains simulated measles outbreak data for the Mandoul region
of Chad. The data is derived from \link[episimdata:moissala_linelist_clean_EN]{episimdata::moissala_linelist_clean_EN} and
includes temporal, demographic, and geographic information for each patient,
with administrative boundary pcodes added for spatial analysis.
}
\examples{
df_ll_measles
}
\keyword{datasets}
