% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{epishiny_tr_names}
\alias{epishiny_tr_names}
\title{Translate the names of a named vector}
\usage{
epishiny_tr_names(x)
}
\arguments{
\item{x}{Named vector with display names.}
}
\value{
A named vector with translated names.
}
\description{
Useful for translating display labels in named vectors such as
\code{date_vars} and \code{group_vars} while preserving column names as values.
}
