% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_place.R
\name{geo_layer}
\alias{geo_layer}
\title{Build a geo layer to be used in the 'place' module}
\usage{
geo_layer(layer_name, sf, name_var, join_by, pop_var = NULL)
}
\arguments{
\item{layer_name}{the name of the geo layer, for example 'State', 'Department', 'Admin2' etc.
If providing multiple layers, layer names must be unique.}

\item{sf}{geographical data of class 'sf' (simple features).}

\item{name_var}{character string of the variable name in \code{sf} containing the names of each geographical feature.}

\item{join_by}{data join specification to join geo layer to a dataset. Should be either a single variable name
present in both datasets or a named vector where the name is the geo layer join variable and the value is the
join variable of the dataset. i.e. \code{c("pcode" = "place_code")} LHS = geo, RHS = data.}

\item{pop_var}{character string of the variable name in \code{sf} containing population data for each feature.
If provided, attack rates will be shown on the map as a choropleth.}
}
\value{
named list of class "epishiny_geo_layer"
}
\description{
Build a geo layer to be used in the 'place' module
}
\examples{
geo_layer(
  layer_name = "Governorate",
  sf = sf_yem$adm1,
  name_var = "adm1_name",
  pop_var = "adm1_pop",
  join_by = c("pcode" = "adm1_pcode")
)
}
