% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{get_epishiny_i18n}
\alias{get_epishiny_i18n}
\title{Get the epishiny Translator object}
\usage{
get_epishiny_i18n()
}
\value{
A \link[shiny.i18n:Translator]{shiny.i18n::Translator} object.
}
\description{
Returns the package-level \link[shiny.i18n:Translator]{shiny.i18n::Translator} instance, creating it on
first use from \href{translations/translation.json}{inst/translations/translation.json}.
}
