% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{init_epishiny_i18n}
\alias{init_epishiny_i18n}
\title{Initialise or reset the epishiny Translator}
\usage{
init_epishiny_i18n(lang = "en")
}
\arguments{
\item{lang}{Two-letter language code (e.g. \code{"en"}, \code{"fr"}).}
}
\value{
The \link[shiny.i18n:Translator]{shiny.i18n::Translator} object (invisibly).
}
\description{
Initialise or reset the epishiny Translator
}
