% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{sf_sle}
\alias{sf_sle}
\title{Sierra Leone Western Area Administrative Boundaries}
\format{
named list of sf objects
}
\usage{
sf_sle
}
\description{
A list of length 2 containing geographic administrative boundary data
for Sierra Leone, stored as simple features (sf) objects.
}
\details{
Each admin level can be joined to the example \code{\link{df_ll_ebola}} dataset with a join by
specification of \code{c("pcode" = "adm3_pcode")} and \code{c("pcode" = "adm4_pcode")}
respectively. These should be passed as the \code{join_by} field in each \code{geo_data}
specification passed to \link{place_ui} and \link{place_server}.
}
\examples{
sf_sle$adm1
sf_sle$adm2
}
\keyword{datasets}
