% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{sf_tcd}
\alias{sf_tcd}
\title{Chad Mandoul Region Administrative Boundaries}
\format{
named list of sf objects
}
\usage{
sf_tcd
}
\description{
A list of length 3 containing geographic administrative boundary data
for the Mandoul region of Chad, stored as simple features (sf) objects.
Includes administrative levels 1, 2, and 4 (Region, Sub-Prefecture and Village/Commune, respectively).
}
\details{
Each admin level can be joined to the example \code{\link{df_ll_measles}} dataset with a join by
specification of \code{c("pcode" = "adm1_pcode")}, \code{c("pcode" = "adm2_pcode")}, and
\code{c("pcode" = "adm4_pcode")} respectively. These should be passed as the \code{join_by}
field in each \code{geo_data} specification passed to \link{place_ui} and \link{place_server}.
}
\examples{
sf_tcd$adm1
sf_tcd$adm2
sf_tcd$adm4
}
\keyword{datasets}
