% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{sf_yem}
\alias{sf_yem}
\title{Yemen Governorate (adm1) and District (adm2) Administrative Boundaries}
\format{
named list of sf objects
}
\usage{
sf_yem
}
\description{
A list of length 2 containing geographic administrative boundary data
for Yemen, stored as simple features (sf) objects.
}
\details{
Each admin level can be joined to the example \code{\link{df_ll}} dataset with a join by
specification of \code{c("pcode" = "adm1_pcode")} and \code{c("pcode" = "adm2_pcode")}
respectively. These should be passed as the \code{join_by} field in each \code{geo_data}
specification passed to \link{place_ui} and \link{place_server}.
}
\examples{
sf_yem$adm1
sf_yem$adm2
}
\keyword{datasets}
