% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{update_epishiny_select}
\alias{update_epishiny_select}
\title{Update a selectInput after a language change}
\usage{
update_epishiny_select(
  session,
  inputId,
  choices_named,
  label = NULL,
  selected = NULL
)
}
\arguments{
\item{session}{Shiny session.}

\item{inputId}{Namespaced input id.}

\item{choices_named}{Named vector with English display names.}

\item{label}{Optional English label key.}

\item{selected}{Optional selected value(s).}
}
\description{
Update a selectInput after a language change
}
