% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{update_epishiny_virtual_select_label}
\alias{update_epishiny_virtual_select_label}
\title{Update a virtualSelectInput label after a language change}
\usage{
update_epishiny_virtual_select_label(session, inputId, label)
}
\arguments{
\item{session}{Shiny session.}

\item{inputId}{Namespaced input id.}

\item{label}{English label key.}
}
\description{
Update a virtualSelectInput label after a language change
}
