% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{use_epishiny_i18n}
\alias{use_epishiny_i18n}
\title{Include shiny.i18n assets in the app UI}
\usage{
use_epishiny_i18n()
}
\value{
A Shiny UI tag (singleton).
}
\description{
Include shiny.i18n assets in the app UI
}
