% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_launch.R
\name{launch_demo_dashboard}
\alias{launch_demo_dashboard}
\title{Launch epishiny demo dashboard}
\usage{
launch_demo_dashboard()
}
\value{
No return value, a shiny app is launched.
}
\description{
See an example of the type of dashboard you can build
using \code{epishiny} modules within a \code{bslib} UI.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  library(epishiny)
  launch_demo_dashboard()
}
}
