% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_time.R
\name{time_ui}
\alias{time_ui}
\alias{time_server}
\title{Time module}
\usage{
time_ui(
  id,
  date_vars,
  group_vars = NULL,
  title = "Time",
  icon = bsicons::bs_icon("bar-chart-line-fill"),
  opts_btn_lab = "options",
  date_lab = "Date axis",
  date_int_lab = "Date interval",
  day_week_month_labs = c("Day", "Week", "Month"),
  groups_lab = "Group data by",
  bar_stacking_lab = "Bar stacking",
  cumul_data_lab = "Show cumulative data?",
  n_lab = "N patients",
  ratio_line_lab = NULL,
  full_screen = TRUE
)

time_server(
  id,
  df_ll,
  date_vars,
  group_vars = NULL,
  y_lab = "Patients",
  n_lab = "N patients",
  ratio_var = NULL,
  ratio_lab = NULL,
  ratio_numer = NULL,
  ratio_denom = NULL,
  filter_info = shiny::reactiveVal()
)
}
\arguments{
\item{id}{Module id. Must be the same in both the UI and server function to link the two.}

\item{date_vars}{named character vector of date variables for the date axis input. Names are used as variable labels.}

\item{group_vars}{Character vector of categorical variable names. If provided, a select input will appear
in the options dropdown allowing for data groups to be visualised as stacked bars on the epicurve.
If named, names are used as variable labels.}

\item{title}{header title for the card.}

\item{icon}{The icon to display next to the title.}

\item{opts_btn_lab}{text label for the dropdown menu button.}

\item{date_lab}{text label for the date variable input.}

\item{date_int_lab}{text label for the date interval input.}

\item{day_week_month_labs}{character vector with text labels for day, week and month, respectively.}

\item{groups_lab}{text label for the grouping variable input.}

\item{bar_stacking_lab}{text label for bar stacking option.}

\item{cumul_data_lab}{text label for cumulative data option.}

\item{n_lab}{The label for the raw count variable.}

\item{ratio_line_lab}{text label for the ratio line input. If not supplied the input is not included.}

\item{full_screen}{Add button to card to with the option to enter full screen mode?}

\item{df_ll}{Data frame or tibble of patient level linelist data. Can be either a shiny reactive or static dataset.}

\item{y_lab}{The label for y-axis of chart.}

\item{ratio_var}{Character string of variable name to use for ratio calculation.}

\item{ratio_lab}{The label to describe the computed ratio i.e. 'CFR' for case fatality ratio.}

\item{ratio_numer}{Value(s) in \code{ratio_var} to be used for the ratio numerator i.e. 'Death'.}

\item{ratio_denom}{Values in \code{ratio_var} to be used for the ratio denominator i.e. \code{c('Death', 'Recovery')}.}

\item{filter_info}{If contained within an app using \code{\link[=filter_server]{filter_server()}}, supply the \code{filter_info} element
returned by that function here as a shiny reactive to add filter information to chart exports.}
}
\value{
the module server function returns any point click event data of the highchart.
see \link[highcharter:hc_add_event_point]{highcharter::hc_add_event_point} for details.
}
\description{
Visualise data over time with an interactive 'epicurve'.
}
